refactor: merge repl/attach subcommands
